- Changed BasicMeshGenerator class to MeshGenerator namespace - Added WebPageRenderer class (currently empty but for future usage -> GUIWebGadget) - Updated SoftPixel Mesh Exporter for Blender 2.63 (supports textures and animation again) - Updated MeshViewer Tool to version 1.2.2 - Fixed SkeletalAnimation problem when several animation joints access same vertices (multi vertex weights)

+++ sources/Base/spTimer.cpp
@@ -112,7 +112,7 @@ u64 Timer::getElapsedMicroseconds()
     // Check for unexpected leaps in the Win32 performance counter
     // (This is caused by unexpected data across the PCI to ISA
     // bridge, aka south bridge. See Microsoft KB274323.)
-    u64 ElapsedTicks = GetTickCount64() - FreqQuery_->StartTick;
+    u64 ElapsedTicks = static_cast<u64>(GetTickCount64() - FreqQuery_->StartTick);
     s64 MSecOff = static_cast<s64>(MSecTicks - ElapsedTicks);
     
     if (MSecOff < -100 || MSecOff > 100)
